fix: 병합 과정 중 변경 및 충돌 난 파일 정리
