+++ apps/web/app/(with-tab)/feed/layout.tsx
@@ -3,17 +3,28 @@
 import { ReactNode } from 'react';
 
 import { AppBar } from '@plog/ui';
+import { usePathname, useRouter } from 'next/navigation';
 
 type FeedLayoutProps = {
   children: ReactNode;
 };
 export default function FeedLayout({ children }: FeedLayoutProps) {
+  const pathname = usePathname();
+  const router = useRouter();
+  const isFeedDetailPage =
+    pathname.startsWith('/feed/') &&
+    pathname.split('/').filter(Boolean).length >= 2;
+
   return (
     <>
-      <header className="fixed inset-x-0 top-0 z-10 mx-auto max-w-layout">
-        <AppBar variant="navigation" title="피드" />
+      <header>
+        <AppBar
+          variant="navigation"
+          title="피드"
+          onBack={isFeedDetailPage ? () => router.push('/feed') : undefined}
+        />
       </header>
-      <div className="pt-[var(--spacing-header)]">{children}</div>
+      {children}
     </>
   );
 }
