--- apps/web/app/feed/page.tsx
@@ -1,3 +1,0 @@
-import { FeedPage } from '@/views/feed';
-
-export default FeedPage;
